requirements.txt
bs4
haystack-ai
Ollama-haystack
markdownify
colorama
tiktoken
simple-term-menu